Moved control of player/star actions into separate script
Now PlayerController is just for movement, and PlayerStarActionController.cs is where the player controls does things with the star
