+++ SoTA - PreProd/Assets/InputActions/PlayerInputActions.inputactions
@@ -49,6 +49,15 @@
                     "processors": "",
                     "interactions": "",
                     "initialStateCheck": false
+                },
+                {
+                    "name": "CarryStarToggle",
+                    "type": "Button",
+                    "id": "98aa978e-5e02-425e-a0d8-08f70b797df0",
+                    "expectedControlType": "",
+                    "processors": "",
+                    "interactions": "",
+                    "initialStateCheck": false
                 }
             ],
             "bindings": [
@@ -192,6 +201,17 @@
                     "processors": "",
                     "groups": "",
                     "action": "LeftMouseRelease",
+                    "isComposite": false,
+                    "isPartOfComposite": false
+                },
+                {
+                    "name": "",
+                    "id": "d937f48d-3c7a-4e63-88b4-7d76d9982b04",
+                    "path": "<Keyboard>/space",
+                    "interactions": "",
+                    "processors": "",
+                    "groups": "",
+                    "action": "CarryStarToggle",
                     "isComposite": false,
                     "isPartOfComposite": false
                 }
