+++ SoTA - PreProd/Assets/Scripts/PlayerStarActionController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c85e22030020bc548815ed96d322c1a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
